Removed repeating code, slight refactoring.
Removed "x_move" and "o_move" methods in favour of one "move" method.
Made slight adjustments to the "last_move" methods in the "BoardData" module and throughout the code to accomodate the changes.

Small refactoring throughout via the use of "then". Most noteably in the "check_win" method.

Two TODOs added.

Also removed the indentation that the whole code had.